Fix: Print correct file/line when a subroutine contains an assert.
Tests such as asserts were implemented as macros containing two parts:

    snow_fail_update(); // Update the globally stored file and line number
    _asserteq(thing 1, thing 2); // Perform a check

This falls down if "thing 1" or "thing 2" contain assertions themselves,
which themselves update the globally stored file and line number. If the
outer assertion then fails, it reports the wrong file and line number.

To fix this, add a struct _snow_explanation that is created in assertion
macros, and is used to update the global file and line number only once
the actual failiure is triggered.